myshop/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
Brotli==1.1.0
celery==5.3.4
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cssselect2==0.7.0
Deprecated==1.2.14
Django==4.2.5
dnspython==2.6.1
eventlet==0.33.3
fonttools==4.49.0
greenlet==3.0.3
html5lib==1.1
idna==3.6
kombu==5.3.5
packaging==23.2
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycparser==2.21
pydyf==0.11.0
pyphen==0.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==4.3.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==10.4.0
tinycss2==1.2.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
weasyprint==59.0
webencodings==0.5.1
wrapt==1.16.0
zopfli==0.2.3
